refactor: switch internal imports to canonical packages
