+++ dfode_kit/data_operations/__init__.py
@@ -22,11 +22,11 @@ __all__ = [
 _ATTRIBUTE_MODULES = {
     "touch_h5": ("dfode_kit.data.io_hdf5", "touch_h5"),
     "get_TPY_from_h5": ("dfode_kit.data.io_hdf5", "get_TPY_from_h5"),
-    "integrate_h5": ("dfode_kit.data_operations.h5_kit", "integrate_h5"),
-    "load_model": ("dfode_kit.data_operations.h5_kit", "load_model"),
-    "nn_integrate": ("dfode_kit.data_operations.h5_kit", "nn_integrate"),
-    "predict_Y": ("dfode_kit.data_operations.h5_kit", "predict_Y"),
-    "calculate_error": ("dfode_kit.data_operations.h5_kit", "calculate_error"),
+    "integrate_h5": ("dfode_kit.data.integration", "integrate_h5"),
+    "load_model": ("dfode_kit.data.integration", "load_model"),
+    "nn_integrate": ("dfode_kit.data.integration", "nn_integrate"),
+    "predict_Y": ("dfode_kit.data.integration", "predict_Y"),
+    "calculate_error": ("dfode_kit.data.integration", "calculate_error"),
     "random_perturb": ("dfode_kit.data_operations.augment_data", "random_perturb"),
     "label_npy": ("dfode_kit.data_operations.label_data", "label_npy"),
     "SCALAR_FIELDS_GROUP": ("dfode_kit.data.contracts", "SCALAR_FIELDS_GROUP"),
